+++ database/migrations/2020_02_27_091820_create_menu_table.php
@@ -18,7 +18,7 @@ class CreateMenuTable extends Migration
             $table->string('nom');
             $table->integer('max_age');
             $table->integer('min_age');
-            $table->integer('calorie');
+            $table->float('calorie');
             $table->integer('type_menu');
             $table->timestamps();
         });
